sensor-graph/java: rename package and fix lint warning
