+++ sensor-graph/app/src/main/AndroidManifest.xml
@@ -19,13 +19,13 @@
 -->
 
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.android.gl2jni">
+    package="com.android.sensorgraph">
     <uses-feature android:glEsVersion="0x00020000" android:required="true"/>
     <application
             android:icon="@mipmap/ic_launcher"
             android:allowBackup="false"
             android:label="@string/gl2jni_activity">
-        <activity android:name=".SensorGraphActivity"
+        <activity android:name="com.android.sensorgraph.SensorGraphActivity"
                 android:theme="@android:style/Theme.NoTitleBar.Fullscreen"
             	android:launchMode="singleTask"
             	android:configChanges="orientation|keyboardHidden">
